+++ vsc-script/src/webview-test1.vsc-script.ts
@@ -1,14 +1,15 @@
 //vsc-script-name: WebView Test > WebView test (test 1)
+//import * as vsc from "./vsc-base-development/vsc-base";
 import * as vsc from "vsc-base";
 import * as vscode from "vscode";
 
 export async function run(path: string, context: vscode.ExtensionContext) {
-   const [postMessage, onMessage, dispose] = vsc.startWebview(context, {
+   const { sendCommand: post, sendSetHTML, onCommand, dispose } = vsc.startWebview(context, {
       title: "Rename",
       showOptions: 2,
       body: `
-      <button onClick="postMessage({command:'show',value:'1'})">Show '1'</button>
-      <button onClick="postMessage({command:'end'})">END</button><br/>
+      <button onClick="sendCommand('show','1')">Show '1'</button>
+      <button onClick="sendCommand('end')">END</button><br/>
       <br><br>
       Search: <input type='text' id='s' onkeyup="postMessage({command:'search',value:this.value})" >
       <br><br>
@@ -16,33 +17,34 @@ export async function run(path: string, context: vscode.ExtensionContext) {
       <br><br>
       <div id='info'>info</div>
     `,
-      onWebviewMessage: (message: any) => {
-         switch (message.command) {
+      onCommand: (command: string, value: any) => {
+         console.log("HERE!!!", command, value)
+         switch (command) {
             case "info":
-               document.getElementById("info")!.innerHTML = message.content;
+               document.getElementById("info")!.innerHTML = value;
                break;
             case "files":
-               document.getElementById("files")!.innerHTML = message.content;
+               document.getElementById("files")!.innerHTML = value;
                break;
          }
       }
    });
-   await onMessage(async (message, resolve) => {
-      switch (message.command) {
+   await onCommand(async (command, value, resolve) => {
+      switch (command) {
          case "show":
-            vsc.showMessage(message.value);
+            vsc.showMessage(value);
             break;
          case "end":
             resolve();
             break;
          case "search":
-            const files = await vsc.findFilePaths(message.value);
-            postMessage({ command: "info", content: 'Found:' + files.length });
-            postMessage({ command: "files", content: files.slice(0, 5).map(x => `<a href='#' onClick="postMessage({command:'open',value:'${x}'})">${x}</a>`) });
+            const files = await vsc.findFilePaths(value);
+            sendSetHTML("#info", 'Found:' + files.length)
+            post("files", files.slice(0, 5).map(x => `<a href='#' onClick="postMessage({command:'open',value:'${x}'})">${x}</a>`));
             break;
          case "open":
-            postMessage({ command: "info", content: `Will open: ${message.value}` });
-            await vsc.open(message.value, 1);
+            sendSetHTML("#info", 'Found:' + `Will open: ${value}`)
+            await vsc.open(value, 1);
             break;
       }
    });